clinical-AI_NW/requirements.txt
fastapi
uvicorn
python-dotenv
openai
faiss-cpu
numpy
pydantic
sentence-transformers
openai